+++ lib/src/nbr/will_overflow_str.c
@@ -20,7 +20,7 @@ bool	will_overflow_str(char* str, size_t n_bits)
 	char*	ptr_charset = strchr(charset, tolower(str[i]));
 	while (str[i] != '\0' && ptr_charset != NULL && number <= 0)
 	{
-		number = radix * number - (ssize_t)(ptr_charset - charset);
+		number = (ssize_t)radix * number - (ssize_t)(ptr_charset - charset);
 		i++;
 		ptr_charset = strchr(charset, tolower(str[i]));
 	}
